requirements.txt
GitPython==3.1.37
GitPython==3.1.43
watchdog==4.0.1
